chore: format code with pre-commit
You never want to commit unformatted code, so it's good that we include
it in the pre-commit hook.
On the other hand, linting may not need to be satisfied in every commit.
For example, one may draft a prototype with a commit and refactor or
apply fixes in later commits.
Thus, linting is enforced with CI.