docs: generate api documentation using swaggo
Although the `docs/` directory can be generated by Swaggo, I have not
included it in the `.gitignore` file. This is to maintain the requirement
of installing the Swaggo command-line tool as a development-only
requirement. This way, individuals who want to quickly try the server
can stil simply run it with `go run main.go`.