fix: database user creation failure
The `init.sql` script was intended to create a user and a database;
however, it did not take effect immediately, resulting in a failure
when the server attempted to connect to the database (Access denied).
By creating these initial user and database with the environment
variables, we are able to resolve this issue.