chore: install tools on demand
Previously, we installed all the tools at once with `make tools`
upfront. However, not all tools may be needed for every task, and this
approach required two separate commands. To streamline the process, we
now check whether a tool is installed when its corresponding target is
executed. If the tool doesn't exist, we install it along with the
execution of the target.