chore: exclude finding go files in "." directories
The primary reason is that we encounter an "access denied" error when
searching in the ".db" directory, which is due to these files being
created in the database container with non-user privileges.
Additionally, it's beneficial not to search in those special directories
because they serve specific purposes and don't need to be included in
the search.