refactor: load environment variables and specify database in main
Since the use of environment variables is a low-level detail, it's
better not to clutter the repositories package with them. This also
provides the advantage of avoiding the need to modify the repositories
package if we choose to set the variables in different ways.
Additionally, specifying the GORM dialect in main allows the
repositories to support multiple types of databases.