feat: record as history only on status change
The `UpdateDocument` function may be used to update the document content
with the status being untouched. In such case, we do not record it.
A database query is issued to retrieve the original status.
On the other hand, the `SetDocumentStatus` function is dedicated to
update the status, thus we believe it always changes the status, and
record it without checking the original status.